requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
psycopg2-binary==2.9.9
redis==5.0.1
httpx==0.25.2
python-dotenv==1.0.0
google-generativeai==0.3.2
celery==5.3.4
email-validator==2.1.0
asyncpg sqlalchemy[asyncio]